+++ src/app/components/Footer.tsx
@@ -12,7 +12,7 @@ export default function Footer() {
           <Link href="https://github.com/Sevih/outerpedia" target="_blank" rel="noopener noreferrer" className="hover:underline inline-flex items-center gap-1">
             <FaGithub /> GitHub
           </Link>
-          <Link href="https://discord.com/invite/outerplaneglobal" target="_blank" rel="noopener noreferrer" className="hover:underline inline-flex items-center gap-1">
+          <Link href="https://discord.com/invite/xUcrH7H8jA" target="_blank" rel="noopener noreferrer" className="hover:underline inline-flex items-center gap-1">
             <FaDiscord /> Outerplane Official Discord
           </Link>
           <Link href="https://discord.gg/keGhVQWsHv" target="_blank" rel="noopener noreferrer" className="hover:underline inline-flex items-center gap-1">