+++ src/app/guides/page.tsx
@@ -1,9 +1,12 @@
-import UnderConstruction from "@/app/components/UnderConstruction";
+import GuideCategoryList from '@/app/components/CategoryCard';
 
-export default function GuidesPage() {
-    return (
-      <main className="min-h-screen bg-gray-900 text-white p-4">
-            <UnderConstruction />
-            </main>
-    );
-  }
+export const dynamic = 'force-static'; // ou 'auto'
+
+export default function GuidesHome() {
+  return (
+    <div className="p-6">
+      <h1 className="text-3xl font-bold mb-6">All Guides</h1>
+      <GuideCategoryList />
+    </div>
+  );
+}